Move contexts into a separate package.
