+++ contexts/vhost.go
@@ -1,4 +1,4 @@
-package vhost
+package contexts
 
 import (
 	"github.com/ironsmile/nedomi/types"
@@ -11,13 +11,13 @@ type vhostContextKey int
 
 const vKey vhostContextKey = 0
 
-// NewContext returns a new Context carrying the supplied vhost pointer.
-func NewContext(ctx context.Context, vhost *types.VirtualHost) context.Context {
+// NewVhostContext returns a new Context carrying the supplied vhost pointer.
+func NewVhostContext(ctx context.Context, vhost *types.VirtualHost) context.Context {
 	return context.WithValue(ctx, vKey, vhost)
 }
 
-// FromContext extracts the vhost.Vhost object, if present.
-func FromContext(ctx context.Context) (*types.VirtualHost, bool) {
+// GetVhost extracts the vhost.Vhost object, if present.
+func GetVhost(ctx context.Context) (*types.VirtualHost, bool) {
 	vhost, ok := ctx.Value(vKey).(*types.VirtualHost)
 	return vhost, ok
 }
